Avoid AnyEvent creating a circular dependency on POE.
POE will not automatically load AnyEvent---a programmer must do that
explicitly.  So the presence of POE::Loop::AnyEvent in a program
indicates a conscious choice by a program's developer.

We must ensure that AnyEvent doesn't subsequently auto-detect and use
POE as its own back end.  If we allow AnyEvent to do this, then a
circular dependency and infinite recursion ensue.

Removing AnyEvent::Impl::POE from the list of available models seems
to be the only way to prevent the circular dependency.  This change
implements that.

If someone knows how to do this without breaking AnyEvent's
encapsulation, advice and patches are always welcome.